+++ src/attr.cpp
@@ -20,7 +20,7 @@ std::ostream& operator<<(std::ostream& o, Attr a)
     case Attr::SORRY: o << "sorry"; break;
     case Attr::LIST: o << "list"; break;
     case Attr::REQUIRES: o << "requires"; break;
-    case Attr::PREMISE_LIST: o << "premise_list"; break;
+    case Attr::PROOF_RULE: o << "proof_rule"; break;
     case Attr::PROGRAM: o << "program"; break;
     case Attr::BINDER: o << "binder"; break;
     case Attr::LET_BINDER: o << "let_binder"; break;