+++ src/state.cpp
@@ -188,10 +188,6 @@ State::State(Options& opts, Stats& stats)
   d_any = Expr(mkExpr(Kind::ANY, {}));
   // self is a distinguished parameter
   d_self = Expr(mkSymbolInternal(Kind::PARAM, "eo::self", d_any));
-  d_conclusion =
-      Expr(mkSymbolInternal(Kind::PARAM, "eo::conclusion", d_boolType));
-  // eo::conclusion is not globally bound, since it can only appear
-  // in :requires.
 }
 
 State::~State() {}
@@ -603,8 +599,6 @@ Expr State::mkSymbol(Kind k, const std::string& name, const Expr& type)
 }
 
 Expr State::mkSelf() const { return d_self; }
-
-Expr State::mkConclusion() const { return d_conclusion; }
 
 Expr State::mkPair(const Expr& t1, const Expr& t2)
 {
@@ -1305,19 +1299,63 @@ Expr State::getProofRule(const std::string& name) const
   return d_null;
 }
 
-bool State::getActualPremises(const ExprValue* rule,
-                              std::vector<Expr>& given,
-                              std::vector<Expr>& actual)
+bool State::getProofRuleArguments(std::vector<Expr>& children,
+                                  Expr& rule,
+                                  Expr& proven,
+                                  std::vector<Expr>& premises,
+                                  std::vector<Expr>& args,
+                                  bool isPop)
 {
-  AppInfo* ainfo = getAppInfo(rule);
-  if (ainfo!=nullptr && ainfo->d_attrCons==Attr::PREMISE_LIST)
+  children.emplace_back(rule.getValue());
+  // arguments first
+  children.insert(children.end(), args.begin(), args.end());
+  AppInfo* ainfo = getAppInfo(rule.getValue());
+  if (ainfo != nullptr)
   {
-    Expr plCons = ainfo->d_attrConsTerm;
+    Attr a = ainfo->d_attrCons;
+    Assert (a==Attr::PROOF_RULE);
+    Expr tupleVal = ainfo->d_attrConsTerm;
+    Assert (tupleVal.getNumChildren()==3);
+    Expr plCons;
+    if (tupleVal[0]!=d_any)
+    {
+      plCons = tupleVal[0];
+    }
+    bool isAssume = tupleVal[1]==d_true;
+    bool isConcExplicit = tupleVal[2]==d_true;
+    if (isConcExplicit)
+    {
+      if (proven.isNull())
+      {
+        // requires a conclusion to be provided
+        return false;
+      }
+      children.push_back(proven);
+    }
+    if (isPop == isAssume)
+    {
+      if (isPop)
+      {
+        std::vector<Expr> as = getCurrentAssumptions();
+        // The size of assumptions should be one, but may contain more
+        // assumptions if e.g. we encountered assume in a nested assumption
+        // scope. Nevertheless, as[0] is always the first assumption in
+        // the assume-push.
+        // push the assumption
+        children.push_back(as[0]);
+      }
+    }
+    else
+    {
+      // using step for a rule requiring an assumption, or step-pop for a rule
+      // not requiring an assumption.
+      return false;
+    }
     if (!plCons.isNull())
     {
       std::vector<Expr> achildren;
       achildren.push_back(plCons);
-      for (Expr& e : given)
+      for (Expr& e : premises)
       {
         // should be proof types
         Expr eproven = d_tc.getType(e);
@@ -1347,14 +1385,26 @@ bool State::getActualPremises(const ExprValue* rule,
         ap = mkExpr(Kind::APPLY, achildren);
       }
       Expr pfap = mkProofType(ap);
-      // TODO: collect operator???
-      // dummy, const term of the given proof type
+      // dummy, "collected" term of the given proof type
       Expr n = mkSymbol(Kind::CONST, "tmp", pfap);
-      actual.push_back(n);
-      return true;
+      children.push_back(n);
+    }
+    else
+    {
+      // otherwise ordinary premises
+      children.insert(children.end(), premises.begin(), premises.end());
     }
   }
-  actual = given;
+  else
+  {
+    // premises after arguments
+    children.insert(children.end(), premises.begin(), premises.end());
+    if (isPop)
+    {
+      // ordinary rule cannot use assumption
+      return false;
+    }
+  }
   return true;
 }